Rename emit module to tracker
