--- src/autometrics/emit/__init__.py
@@ -1,1 +1,0 @@
-from .emit import *
